# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FitAI API
  description: |
    API completa para FitAI - Aplicación de fitness con IA para el mercado chileno.
    
    ## Autenticación
    La mayoría de endpoints requieren autenticación mediante Clerk. Include el token en el header:
    ```
    Authorization: Bearer YOUR_CLERK_TOKEN
    ```
    
    ## Planes de Suscripción
    - **Free**: Funcionalidades básicas
    - **Premium**: IA avanzada, analytics, social features ($7,990 CLP/mes)
    - **Pro**: Todas las funcionalidades premium + análisis avanzado ($14,990 CLP/mes)
    
    ## Rate Limits
    - API general: 1000 requests/hour
    - IA features:
      - Free: 3 rutinas, 10 consejos por día
      - Premium: 15 rutinas, 50 consejos por día  
      - Pro: 50 rutinas, 200 consejos por día
  version: 1.0.0
  contact:
    name: FitAI Support
    email: support@fitai.cl
  license:
    name: Proprietary
servers:
  - url: https://api.getfitia.com
    description: Production server
  - url: https://fitai-api.sweetspot-627.workers.dev
    description: Development server

# Security schemes
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk authentication token
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "ValidationError"
        message:
          type: string
          example: "Invalid input data"
        path:
          type: string
          example: "/api/v1/users/me"
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          example: "user_123"
        clerkUserId:
          type: string
          example: "clerk_456"
        email:
          type: string
          example: "usuario@example.com"
        name:
          type: string
          example: "Juan Pérez"
        plan:
          type: string
          enum: [free, premium, pro]
          example: "premium"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        profile:
          $ref: '#/components/schemas/UserProfileData'
        stats:
          $ref: '#/components/schemas/UserStats'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserProfileData:
      type: object
      properties:
        goals:
          type: array
          items:
            type: string
          example: ["muscle_gain", "strength"]
        experienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
          example: "intermediate"
        availableDays:
          type: integer
          minimum: 1
          maximum: 7
          example: 4
        height:
          type: number
          example: 175
        weight:
          type: number
          example: 70
        age:
          type: integer
          example: 25
        equipment:
          type: array
          items:
            type: string
          example: ["dumbbell", "barbell"]
        workoutLocation:
          type: string
          example: "gym"
        injuries:
          type: array
          items:
            type: string
          example: []
    UserStats:
      type: object
      properties:
        workoutsCompleted:
          type: integer
          example: 45
        currentStreak:
          type: integer
          example: 7
        prsSet:
          type: integer
          example: 12
    
    # Exercise schemas
...
security:
  - ClerkAuth: []
paths:
...
  /api/v1/users/me:
    get:
      tags:
        - Users
      summary: Get Current User Profile
      description: Obtiene el perfil completo del usuario actual con estadísticas
      responses:
        '200':
          description: Perfil del usuario obtenido exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '401':
          description: Usuario no autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Perfil de usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update User Profile
      description: Actualiza el perfil del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Juan Pérez Actualizado"
                profile:
                  $ref: '#/components/schemas/UserProfileData'
      responses:
        '200':
          description: Perfil actualizado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      message:
                        example: "Perfil actualizado exitosamente"
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete User Account
      description: Elimina la cuenta del usuario (soft delete)
      responses:
        '200':
          description: Cuenta eliminada exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      message:
                        example: "Cuenta eliminada exitosamente"
  /api/v1/users/me/progress:
    get:
      tags:
        - Users
      summary: Get User Progress Metrics
      description: Obtiene métricas de progreso del usuario
      parameters:
        - name: period
          in: query
          description: Período de tiempo para las métricas
          schema:
            type: string
            enum: [last_7_days, last_30_days, last_90_days]
            default: last_30_days
      responses:
        '200':
          description: Métricas de progreso obtenidas exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          period:
                            type: string
                            example: "last_30_days"
                          metrics:
                            type: object
                            properties:
                              totalWorkouts:
                                type: integer
                                example: 12
                              avgWorkoutDuration:
                                type: integer
                                example: 65
                              totalVolume:
                                type: number
                                example: 8500
                              avgRPE:
                                type: string
                                example: "7.2"
                              consistencyScore:
                                type: integer
                                example: 87
                          personalRecords:
                            type: array
                            items:
                              type: object
                              properties:
                                exerciseId:
                                  type: string
                                exerciseName:
                                  type: string
                                recordType:
                                  type: string
                                value:
                                  type: number
                                unit:
                                  type: string
                                achievedAt:
                                  type: string
                                  format: date-time
                          weeklyFrequency:
                            type: object
                            properties:
                              target:
                                type: integer
                              actual:
                                type: string
                              days:
                                type: array
                                items:
                                  type: boolean
  /api/v1/users/me/preferences:
    put:
      tags:
        - Users
      summary: Update User Preferences
      description: Actualiza las preferencias del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                language:
                  type: string
                  example: "es"
                units:
                  type: string
                  enum: [metric, imperial]
                  example: "metric"
                notifications:
                  type: object
                  properties:
                    workouts:
                      type: boolean
                    achievements:
                      type: boolean
                    social:
                      type: boolean
      responses:
        '200':
          description: Preferencias actualizadas exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      message:
                        example: "Preferencias actualizadas exitosamente"

  # Exercises endpoints
...